pyproject.toml
[tool.poetry]
name = "hwdjango"
version = "0.1.0"
description = ""
authors = ["Ivan Lysogor <ilysogor@yandex-team.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
psycopg2-binary = "^2.8.6"
django-debug-toolbar = "^3.2.1"
mixer = "^7.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
